requirements.txt
Django==1.6
MySQL-python==1.2.5
Pillow==2.7.0
South==1.0.2
Twisted==13.2.0
altgraph==0.12
bdist-mpkg==0.5.0
bonjour-py==0.3
django-classy-tags==0.6.1
django-paypal==0.2
django-polymorphic==0.6.1
django-shop==0.2.0
django-shop-ipayment==0.0.3
django-shop-paypal==0.0.1
django-shop-simplevariations==0.0.4
jsonfield==1.0.3
macholib==1.7
matplotlib==1.4.2
mock==1.0.1
modulegraph==0.12
nose==1.3.4
numpy==1.9.1
py==1.4.26
py2app==0.9
pyOpenSSL==0.14
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyparsing==2.0.3
python-dateutil==2.4.0
pytz==2014.10
scipy==0.14.1
six==1.9.0
tox==1.9.0
virtualenv==12.0.7
wsgiref==0.1.2
xattr==0.6.4
zope.interface==4.1.2
